+++ settings.gradle
@@ -2,4 +2,4 @@ include 'quartermaster'
 include 'infiniboard-app'
 include 'harvester'
 include 'test-base'
-include 'overseer'
+include 'packrat'
